+++ crates/biome_cli/src/diagnostics.rs
@@ -716,7 +716,7 @@ mod test {
     fn termination_diagnostic_size() {
         assert_eq!(
             std::mem::size_of::<CliDiagnostic>(),
-            104,
+            96,
             "you successfully decreased the size of the diagnostic!"
         )
     }