+++ crates/biome_configuration/src/javascript/formatter.rs
@@ -1,7 +1,7 @@
 use crate::PlainIndentStyle;
 use crate::{deserialize_line_width, serialize_line_width};
 use biome_deserialize_macros::{Deserializable, Merge, Partial};
-use biome_formatter::{LineEnding, LineWidth, QuoteStyle};
+use biome_formatter::{AttributePosition, LineEnding, LineWidth, QuoteStyle};
 use biome_js_formatter::context::trailing_comma::TrailingComma;
 use biome_js_formatter::context::{ArrowParentheses, QuoteProperties, Semicolons};
 use bpaf::Bpaf;
@@ -87,6 +87,15 @@ pub struct JavascriptFormatter {
     /// The type of quotes used in JavaScript code. Defaults to double.
     #[partial(bpaf(long("quote-style"), argument("double|single"), optional))]
     pub quote_style: QuoteStyle,
+
+    // it's also a top-level configurable property.
+    /// The attribute position style in JavaScript code. Defaults to auto.
+    #[partial(bpaf(
+        long("javascript-attribute-position"),
+        argument("multiline|auto"),
+        optional
+    ))]
+    pub attribute_position: AttributePosition,
 }
 
 impl Default for JavascriptFormatter {
@@ -106,6 +115,7 @@ impl Default for JavascriptFormatter {
             line_ending: Default::default(),
             line_width: Default::default(),
             quote_style: Default::default(),
+            attribute_position: Default::default(),
         }
     }
 }