+++ crates/biome_css_parser/src/syntax/at_rule/mod.rs
@@ -5,6 +5,7 @@ mod counter_style;
 mod document;
 mod feature;
 mod font_face;
+mod font_feature_values;
 mod font_palette_values;
 mod import;
 mod keyframes;
@@ -28,6 +29,9 @@ use crate::syntax::at_rule::counter_style::{
 };
 use crate::syntax::at_rule::document::{is_at_document_at_rule, parse_document_at_rule};
 use crate::syntax::at_rule::font_face::{is_at_font_face_at_rule, parse_font_face_at_rule};
+use crate::syntax::at_rule::font_feature_values::{
+    is_at_font_feature_values_at_rule, parse_font_feature_values_at_rule,
+};
 use crate::syntax::at_rule::font_palette_values::{
     is_at_font_palette_values_at_rule, parse_font_palette_values_at_rule,
 };
@@ -87,6 +91,8 @@ pub(crate) fn parse_any_at_rule(p: &mut CssParser) -> ParsedSyntax {
         parse_container_at_rule(p)
     } else if is_at_font_face_at_rule(p) {
         parse_font_face_at_rule(p)
+    } else if is_at_font_feature_values_at_rule(p) {
+        parse_font_feature_values_at_rule(p)
     } else if is_at_font_palette_values_at_rule(p) {
         parse_font_palette_values_at_rule(p)
     } else if is_at_media_at_rule(p) {