+++ crates/biome_js_formatter/tests/specs/js/module/expression/unary_expression.js
@@ -1,4 +1,7 @@
 delete  a.a
+delete  (a.a)
+delete  console.log(1)
+delete  (() => {})
 void    b
 typeof  c
 x = +1