+++ CHANGELOG.md
@@ -12,7 +12,22 @@ Read our [guidelines for writing a good changelog entry](https://github.com/biom
 
 ### Analyzer
 
+#### Bug fixed
+
+- Fix [#1748](https://github.com/biomejs/biome/issues/1748). Now for the following case we won't provide an unsafe fix for the `noNonNullAssertion` rule:
+
+  ```ts
+  x[y.z!];
+  ```
+
+  Contributed by @ah-yu
+
 ### CLI
+
+#### New features
+
+- Add a new command `biome migrate prettier`. The command will read the file `.prettierrc`/`prettier.json` and `.prettierignore` and map its configuration to Biome's one.
+  Due to the different nature of `.prettierignore` globs and Biome's globs, it's **highly** advised to make sure that those still work under Biome.
 
 #### Bug fixes
 
@@ -27,6 +42,17 @@ Read our [guidelines for writing a good changelog entry](https://github.com/biom
 - Fix [#1173](https://github.com/biomejs/biome/issues/1173). Fix the formatting of a single instruction with commented in a control flow body to ensure consistency. Contributed by @mdm317
 
 - Fix overriding of `javascript.globals`. Contributed by @arendjr
+- Fix a bug where syntax rules weren't run when pulling the diagnostics. Now Biome will emit more parsing diagnostics, e.g.
+  ```
+  check.js:1:17 parse/noDuplicatePrivateClassMembers ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+
+    × Duplicate private class member "#foo"
+
+    > 1 │ class A { #foo; #foo }
+        │                 ^^^^
+
+  ```
+  Contributed by @ematipico
 
 ### Configuration
 
@@ -35,6 +61,19 @@ Read our [guidelines for writing a good changelog entry](https://github.com/biom
 #### Bug fixes
 
 - Fix a regression where ignored files where formatted in the editor. Contributed by @ematipico
+- Fix a bug where syntax rules weren't run when pulling the diagnostics. Now Biome will emit more parsing diagnostics, e.g.
+  ```
+  check.js:1:17 parse/noDuplicatePrivateClassMembers ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+
+    × Duplicate private class member "#foo"
+
+    > 1 │ class A { #foo; #foo }
+        │                 ^^^^
+
+  ```
+  Contributed by @ematipico
+
+
 
 ### Formatter
 
@@ -42,6 +81,17 @@ Read our [guidelines for writing a good changelog entry](https://github.com/biom
 
 - Biome now allows to format the `package.json` file. This is now the default behaviour and users can remove their workarounds.
   If you rely on other tools to format `package.json`, you'll have to ignore it via configuration. Contributed by @pattrickrice
+- New formatter option `attributePosition` that have similar behavior as Prettier `singleAttributePerLine` [#1706](https://github.com/biomejs/biome/issues/1706). Contributed by @octoshikari
+- Add partial for `.astro` files. Biome is able to format the frontmatter of the Astro files. Contributed by @ematipico
+
+  ```diff
+  ---
+  - statement ( );
+  + statement();
+  ---
+
+  <div></div>
+  ```
 
 #### Bug fixes
 
@@ -128,7 +178,38 @@ Read our [guidelines for writing a good changelog entry](https://github.com/biom
 
 - Fix [#1704](https://github.com/biomejs/biome/issues/1704). Convert `/` to escaped slash `\/` to avoid parsing error in the result of autofix. Contributed by @togami2864
 
+- Fix[#1697](https://github.com/biomejs/biome/issues/1697). Preserve leading trivia in autofix of suppression rules. Contributed by @togami2864
+
+- Fix[#603](https://github.com/biomejs/biome/issues/603). Trim trailing whitespace to avoid double insertion. Contributed by @togami2864
+
 ### Parser
+
+#### Bug fixes
+- Fix [#1728](https://github.com/biomejs/biome/issues/1728). Correctly parse the global declaration when the `{` token is on the line following the `global` keyword.
+
+  Now the following code is correctly parsed:
+
+  ```ts
+  declare global
+  { }
+
+  declare module foo {
+    global
+    { }
+  }
+  ```
+
+  Contributed by @ah-yu
+
+- Fix [#1730](https://github.com/biomejs/biome/issues/1730). Correctly parse `delete` expressions with operands that are not simple member expressions.
+
+  ```js
+  delete(a.b);
+  delete console.log(1);
+  delete(() => {});
+  ```
+
+  Contributed by @printfn
 
 ## 1.5.3 (2024-01-22)